[#18]refactor:types 디렉토리 경로 수정
